@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "unique_follow_up_questions"
-version = "0.0.4"
+version = "0.0.5"
 description = ""
 authors = [
     "Martin Fadler <martin.fadler@unique.ch>",
@@ -12,7 +12,7 @@ readme = ["README.md", "CHANGELOG.md"]
 license = "Proprietary"
 
 [tool.poetry.dependencies]
-python = "^3.11"
+python = "^3.12"
 typing-extensions = "^4.9.0"
 pydantic = "^2.8.2"
 pydantic-settings = "^2.10.1"
@@ -21,11 +21,11 @@ pytest = "^8.4.1"
 jinja2 = "^3.1.6"
 pillow = "^10.4.0"
 unique-sdk = "^0.10.0"
-unique-toolkit = "^0.8.18"
+unique-toolkit = "^0.9.0"
 
 
 [tool.poetry.group.dev.dependencies]
-python = "^3.11"
+python = "^3.12"
 typing-extensions = "^4.9.0"
 pydantic = "^2.8.2"
 pydantic-settings = "^2.10.1"
