@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "unique_stock_ticker"
-version = "0.0.5"
+version = "0.0.6"
 description = ""
 authors = [
     "Martin Fadler <martin.fadler@unique.ch>",
@@ -12,13 +12,13 @@ readme = ["README.md", "CHANGELOG.md"]
 license = "Proprietary"
 
 [tool.poetry.dependencies]
-python = "^3.11"
+python = "^3.12"
 typing-extensions = "^4.9.0"
 pydantic = "^2.8.2"
 pydantic-settings = "^2.10.1"
 python-dotenv = "^1.0.1"
 unique-sdk = "^0.10.0"
-unique-toolkit = "^0.8.18"
+unique-toolkit = "^0.9.0"
 plotly = "^5.24.1"
 numerize = "^0.12"
 pytest = "^8.4.1"
@@ -27,7 +27,7 @@ pyopenssl = "^25.1.0"
 
 
 [tool.poetry.group.dev.dependencies]
-python = "^3.11"
+python = "^3.12"
 typing-extensions = "^4.9.0"
 pydantic = "^2.8.2"
 pydantic-settings = "^2.10.1"
