@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "unique_internal_search"
-version = "0.0.1"
+version = "0.0.7"
 description = ""
 authors = [
     "Martin Fadler <martin.fadler@unique.ch>",
@@ -12,18 +12,19 @@ readme = ["README.md", "CHANGELOG.md"]
 license = "Proprietary"
 
 [tool.poetry.dependencies]
-python = "^3.11"
+python = "^3.12"
 typing-extensions = "^4.9.0"
 pydantic = "^2.8.2"
 pydantic-settings = "^2.10.1"
 python-dotenv = "^1.0.1"
 pytest = "^8.4.1"
+pillow = "^10.4.0"
 unique-sdk = "^0.10.0"
-unique-toolkit = "^0.8.18"
+unique-toolkit = "^0.9.0"
 
 
 [tool.poetry.group.dev.dependencies]
-python = "^3.11"
+python = "^3.12"
 typing-extensions = "^4.9.0"
 pydantic = "^2.8.2"
 pydantic-settings = "^2.10.1"
