@@ -1,18 +1,20 @@
 [tool.poetry]
 name = "unique_toolkit"
-version = "0.8.22"
+version = "0.9.1"
 description = ""
 authors = [
+    "Cedric Klinkert <cedric.klinkert@unique.ch>",
     "Martin Fadler <martin.fadler@unique.ch>",
     "Sadique Sheik <sadique@unique.ch>",
     "Fabian Schläpfer <fabian@unique.ch>",
     "Pascal Hauri <pascal@unique.ch>",
+    "Ejiroghene Oki-Peter<oki-peter.ejiroghene@unique.ai>",
 ]
 readme = ["README.md", "CHANGELOG.md"]
 license = "Proprietary"
 
 [tool.poetry.dependencies]
-python = "^3.11"
+python = "^3.12"
 typing-extensions = "^4.9.0"
 pydantic = "^2.8.2"
 pyhumps = "^3.8.0"
@@ -20,10 +22,15 @@ numpy = "^1.26.4"
 python-dotenv = "^1.0.1"
 regex = "^2024.5.15"
 tiktoken = "^0.7.0"
-unique-sdk = "^0.10.0"
 pydantic-settings = "^2.10.1"
 sseclient = "^0.0.27"
 openai = "^1.99.9"
+platformdirs = "^4.0.0"
+pillow = "^10.4.0"
+unique-sdk = "^0.10.0"
+jambo = "^0.1.2"
+mkdocs-multirepo-plugin = "^0.8.3"
+mkdocs-mermaid2-plugin = "^1.2.2"
 
 [tool.poetry.group.langchain]
 optional = true
@@ -42,8 +49,19 @@ pytest-cov = "^4.1.0"
 pre-commit = "^3.7.1"
 pytest-asyncio = "^0.23.8"
 pytest-mock = "^3.14.0"
-unique-sdk = { path = "../unique_sdk" }
+platformdirs = "^4.0.0"
+pillow = "^10.4.0"
 ipykernel = "^6.29.5"
+# unique-sdk = { path = "../unique_sdk" }
+mkdocs = "^1.6.1"
+mkdocs-material = "^9.6.11"
+mkdocs-include-markdown-plugin = "^7.1.5"
+mkdocs-callouts = "^1.16.0"
+mkdocstrings-python = "^1.16.10"
+entangled-cli = "^2.1.13"
+mkdocs-entangled-plugin = "^0.4.0"
+isort = "^6.0.1"
+mkdocs-codeinclude-plugin = "^0.2.1"
 
 [build-system]
 requires = ["poetry-core"]
@@ -51,6 +69,8 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.ruff]
 target-version = "py311"
+exclude = [".venv", "docs"]
+
 
 [tool.ruff.lint]
 extend-select = ["I"]
