@@ -5,8 +5,8 @@ from unique_toolkit._common.default_language_model import (
     DEFAULT_GPT_4o_STRUCTURED_OUTPUT,
 )
 from unique_toolkit._common.validators import LMI
+from unique_toolkit.agentic.tools.config import get_configuration_dict
 from unique_toolkit.language_model.infos import LanguageModelInfo, ModelCapabilities
-from unique_toolkit.tools.config import get_configuration_dict
 
 from unique_follow_up_questions.prompts.params import (
     FOLLOW_UP_QUESTION_SYSTEM_PROMPT_TEMPLATE,
