@@ -10,8 +10,8 @@ from pydantic import (
 )
 from typing_extensions import deprecated
 from unique_toolkit._common.validators import LMI
+from unique_toolkit.agentic.tools.config import get_configuration_dict
 from unique_toolkit.language_model.infos import LanguageModelInfo, LanguageModelName
-from unique_toolkit.tools.config import get_configuration_dict
 
 from unique_stock_ticker.detection.config import StockTickerDetectionConfig
 from unique_stock_ticker.detection.memory import StockTickerMemoryConfig
