@@ -1,13 +1,13 @@
 from typing import Self
 
 from pydantic import BaseModel
-from unique_toolkit.short_term_memory.persistent_short_term_memory_manager import (
+from unique_toolkit.agentic.short_term_memory_manager.persistent_short_term_memory_manager import (
     PersistentShortMemoryManager,
 )
-from unique_toolkit.short_term_memory.service import ShortTermMemoryService
-from unique_toolkit.tools.config import (
+from unique_toolkit.agentic.tools.config import (
     get_configuration_dict,
 )
+from unique_toolkit.short_term_memory.service import ShortTermMemoryService
 
 from unique_stock_ticker.detection.schema import StockTicker
 
