@@ -3,7 +3,7 @@ from enum import StrEnum
 from typing import Generic, TypeVar
 
 from pydantic import BaseModel
-from unique_toolkit.tools.config import get_configuration_dict
+from unique_toolkit.agentic.tools.config import get_configuration_dict
 
 from unique_stock_ticker.plot.backend.base.schema import (
     StockHistoryPlotPayload,
