@@ -2,7 +2,12 @@
 from unique_toolkit.chat import ChatService
 from unique_toolkit.content import ContentService
 from unique_toolkit.embedding import EmbeddingService
-from unique_toolkit.language_model import LanguageModelMessages, LanguageModelService
+from unique_toolkit.language_model import (
+    LanguageModelMessages,
+    LanguageModelName,
+    LanguageModelService,
+    LanguageModelToolDescription,
+)
 from unique_toolkit.short_term_memory import ShortTermMemoryService
 
 # You can add other classes you frequently use here as well
@@ -10,6 +15,8 @@ from unique_toolkit.short_term_memory import ShortTermMemoryService
 __all__ = [
     "LanguageModelService",
     "LanguageModelMessages",
+    "LanguageModelName",
+    "LanguageModelToolDescription",
     "ChatService",
     "ContentService",
     "EmbeddingService",
